requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
pvporcupine
sounddevice
pyaudio
azure.cognitiveservices.speech
flask_socketio
simple-websocket
pyyaml
wit
openai
elevenlabs
webcolors
python-engineio
google-cloud-dialogflow
googlemaps
Eel